+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Upgrade the `org.realityforge.grim` artifacts to version `0.05`.
 * Upgrade the `au.com.stocksoftware.idea.codestyle` artifact to version `1.17`.
+* Migrate from Elemental2 to Akasha when interacting with the Browser API.
 
 ### [v0.21](https://github.com/arez/arez-persist/tree/v0.21) (2020-12-19) · [Full Changelog](https://github.com/arez/arez-persist/compare/v0.20...v0.21)